requirements.txt
torch==2.4.0
torch_geometric==2.5.3
biopython==1.84
biotite==1.0.0
